Moving out tests from mermaid.spec.js
